Adds support for environment based dependency injection (#94)
Add environment configuration for DI

Env class is now available to configure development and production
environments. New environment strings can be added as static constant members
to the class.

setupLocator() takes in the environment string as an argument and
dependencies can be annotated accordingly so that correct ones can be
injected based on the environment being used.